# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Complete Ecormmerce server.",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/admin/generate-coupens/": {
            "post": {
                "description": "admin userdata",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "admin-coupen",
                "operationId": "admin-userdata-coupen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "coupen code",
                        "name": "coupen_code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "discount of the coupen",
                        "name": "discount",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "quantity code",
                        "name": "quantity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "validity code",
                        "name": "validity",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/admin/userdata/": {
            "get": {
                "description": "admin userdata",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin User Handler"
                ],
                "summary": "admin-userdata",
                "operationId": "admin-userdata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user name",
                        "name": "search",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/cancelorder": {
            "get": {
                "description": "cancel order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "calcel order",
                "operationId": "cancel orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the user",
                        "name": "user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order id of the product ",
                        "name": "orderid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/returnorder": {
            "get": {
                "description": "return order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "return order",
                "operationId": "return orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the user",
                        "name": "user",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/vieworder": {
            "get": {
                "description": "view orders",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "viewOrders",
                "operationId": "viewOrders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the user",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
...
}